Mod:templates

+++ AtCoder/abc/abc999_template/c.rs
@@ -2,11 +2,12 @@
 #![allow(non_snake_case)]
 #![allow(dead_code)]
 #![allow(non_upper_case_globals)]
+#![allow(unused_variables)]
 
 use proconio::marker::*;
 use proconio::*;
 use std::collections::{BinaryHeap, HashMap, HashSet};
-use std::{f64, i64};
+use std::{f64, i64, usize};
 
 const inf: i64 = i64::MAX / 10;
 const md: i64 = 1_000_000_007;
@@ -14,6 +15,6 @@ const eps: f64 = 1e-10;
 
 fn main() {
     input! {
-        N:i64,
+        N:usize,
     }
 }
